making torch models compatible

+++ models.py
@@ -11,7 +11,7 @@ class SiameseNet(nn.Module):
         super(SiameseNet, self).__init__()
         if mask:
             mask = np.tril(np.ones(256), k=-3)+np.triu(np.ones(256), k=3)
-            self.mask = nn.Parameter(torch.tensor([mask]), requires_grad = False)
+            self.mask = nn.Parameter(torch.tensor([mask], dtype=torch.int32), requires_grad = False)
     def mask_data(self, x):
         if hasattr(self, "mask"): x=torch.mul(self.mask, x)
         return x
